--- src/Gateway/Gateway.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace Phlow\Gateway;
-
-use Phlow\Node\Node;
-
-/**
- * Interface Gateway
- * A gateway denotes forking and merging of workflow paths.
- * @package Phlow\Gateway
- */
-interface Gateway extends Node
-{
-
-}
