+++ src/Node/Conditional.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Phlow\Node;
+
+/**
+ * Interface Node
+ * A gateway denotes forking and merging of workflow paths.
+ * @package Phlow\Node
+ */
+interface Conditional extends Node
+{
+
+}
